requirements.txt
google-auth>=2.25.2
requests>=2.31.0
slack-bolt>=1.18.1
